% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrRF-Class.R
\docType{class}
\name{cbrRF}
\alias{cbrRF}
\title{RandomForest Proximity}
\format{An \code{\link{R6Class}} generator object}
\usage{
cbrRF
}
\arguments{
\item{learning:}{data set for learning the RF model}

\item{queryData:}{Query data set. For each case in the verum data, we are 
looking for the k (=1,…,l) similar cases. Learning and verum data set need 
the same structure (variable names and scales)}

\item{learnVars}{(Default: all variables except endPoint): A character vector 
variable names. This variables are used for learning the model. Do not 
include time2event and event variable here.}

\item{endPoint}{(Default: c("Time2Event", "Event")): A character vector of 
length two. The first elements contains the variable name of the time 2 event 
variable and the second the name of the event variable.}

\item{impute}{(Default: FALSE): Missing value imputation.}
}
\description{
This class uses the proximity matrix of the random survival forest algorithm 
as a similarity matrix (sqrt(1 - proximity matrix)) of learning and verum 
cases. By default all cases with at least one missing values are dropped 
from learning, calculating the distance matrix, and searching for similar
cases.
}
\keyword{Cox}
\keyword{Model}

