% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrValidate-Class.R
\docType{class}
\name{cbrValidate}
\alias{cbrValidate}
\title{R6 Validation Class for case based reasoning}
\format{An \code{\link{R6Class}} generator object}
\usage{
cbrValidate
}
\description{
R6 Validation Class for case based reasoning
}
\keyword{validation}

