% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfDistance.R
\name{depthMatrixRanger}
\alias{depthMatrixRanger}
\title{Get mean node dist length}
\usage{
depthMatrixRanger(x, rf)
}
\description{
Get mean node dist length
}
\examples{
\dontrun{
require(ranger)
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
depthMatrixRanger(iris[, -5], rf)
}

}

