% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance calculation}
\usage{
distance(x, y = NULL, w = NULL, method = "euclidian", p = 2)
}
\arguments{
\item{x}{a new dataset}

\item{rf}{\code{ranger} object}
}
\value{
a \code{dist} or \code{matrix} object
}
\description{
Distance calculation
}
\examples{
\dontrun{
require(ranger)
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
terminalNodeIdsRanger(iris[, -5], rf)
}

}

