% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{orderCPP}
\alias{orderCPP}
\title{C++ Version for ordering a matrix or an vector}
\usage{
orderCPP(x, decreasing = F, k = 0)
}
\arguments{
\item{x}{a \code{matrix} or \code{vector}}

\item{decreasing}{Desceding or ascending ordering}

\item{k}{returns k first elements (default k = 0: all elements)}
}
\description{
C++ Version for ordering a matrix or an vector
}
