% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfDistance.R
\name{rangerTreeAsMat}
\alias{rangerTreeAsMat}
\title{Transform trees of a \code{ranger}-object to a matrix}
\usage{
rangerTreeAsMat(rf)
}
\value{
a \code{matrix} object with 
Column 1: tree ID
Column 2: node ID
Column 3: child node ID 1
Column 4: child node ID 2
}
\description{
Transform trees of a \code{ranger}-object to a matrix
}
\examples{
\dontrun{
require(ranger)
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
rangerTreeAsMat(rf)
}
}

