% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrSimCases-Class.R
\docType{class}
\name{simCases}
\alias{simCases}
\title{R6 reference class for getting the similar cases}
\format{An \code{\link{R6Class}} generator object}
\usage{
simCases
}
\description{
R6 reference class for getting the similar cases
}
\keyword{Cox}
\keyword{Model}

