\docType{package}
\name{similarity-package}
\alias{similarity-package}
\title{Similarity measurement for binary matrices.}
\description{
  This package provides functions for similarity
  measurement for binary matrices.
}
\details{
  \tabular{ll}{ Package: \tab similarity \cr Version: \tab
  0.2\cr Date: \tab 2012-09-01\cr License: \tab GPL (>=
  3)\cr URL: \tab http://www.github.com/sgibb/similarity/
  \cr }
}
\author{
  Sebastian Gibb <\email{mail@sebastiangibb.de}>
}
\references{
  \url{http://www.github.com/sgibb/similarity/}
}
\keyword{package}

