% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfDistance.R
\name{terminalNodeDistance}
\alias{terminalNodeDistance}
\title{Terminal node distance for each tree and terminal}
\usage{
terminalNodeDistance(rf)
}
\arguments{
\item{rf}{\code{ranger} object}
}
\value{
a \code{matrix} object with pairwise terminal node edge length
}
\description{
first two columns are terminal node IDs; If an ID pair do not appear in a 
tree -1 is inserted
}
\examples{
\dontrun{
require(ranger)
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
terminalNodeDistance(rf)
}

}
