% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfDistance.R
\name{depthMatrixRanger}
\alias{depthMatrixRanger}
\title{Get depth distance matrix}
\usage{
depthMatrixRanger(x, y = NULL, rf)
}
\arguments{
\item{x}{a new dataset}

\item{y}{a new dataset}

\item{rf}{\code{ranger} object}
}
\description{
Get depth distance matrix
}
\examples{
\dontrun{
require(ranger)
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
depthMatrixRanger(x=iris[, -5], rf=rf)
}

}

