% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance calculation}
\usage{
distance(x, y = NULL, method = "euclidian", p = 2)
}
\arguments{
\item{x}{a new dataset}

\item{y}{a second new dataset}

\item{method}{a distance method (available are: euclidian, manhattan, maximum, minkowski, and cosine)}

\item{p}{parameter for Minkowski distance}
}
\value{
a \code{dist} or \code{matrix} object
}
\description{
Distance calculation
}
\examples{
\dontrun{
distance(iris[, -5])
}

}
