% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity.R
\docType{package}
\name{similarity}
\alias{similarity}
\alias{similarity-package}
\title{similarity is a package for similarity calculation}
\description{
similarity is a package for similarity calculation
}
